refactor(python): Explicitly mark `dependencies` module as private
