+++ py-polars/polars/_dependencies.py
@@ -256,7 +256,7 @@ def import_optional(
 
     Examples
     --------
-    >>> from polars.dependencies import import_optional
+    >>> from polars._dependencies import import_optional
     >>> import_optional(
     ...     "definitely_a_real_module",
     ...     err_prefix="super-important package",
