+++ py-polars/polars/_utils/async_.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, Awaitable, Generator, Generic, TypeVar
 
+from polars._dependencies import _GEVENT_AVAILABLE
 from polars._utils.wrap import wrap_df
-from polars.dependencies import _GEVENT_AVAILABLE
 
 if TYPE_CHECKING:
     from asyncio.futures import Future
