+++ py-polars/polars/_utils/construction/series.py
@@ -15,6 +15,14 @@ from typing import (
 
 import polars._reexport as pl
 import polars._utils.construction as plc
+from polars._dependencies import (
+    _PYARROW_AVAILABLE,
+    _check_for_numpy,
+    dataclasses,
+)
+from polars._dependencies import numpy as np
+from polars._dependencies import pandas as pd
+from polars._dependencies import pyarrow as pa
 from polars._utils.construction.utils import (
     get_first_non_none,
     is_namedtuple,
@@ -52,21 +60,13 @@ from polars.datatypes.constructor import (
     polars_type_to_constructor,
     py_type_to_constructor,
 )
-from polars.dependencies import (
-    _PYARROW_AVAILABLE,
-    _check_for_numpy,
-    dataclasses,
-)
-from polars.dependencies import numpy as np
-from polars.dependencies import pandas as pd
-from polars.dependencies import pyarrow as pa
 
 with contextlib.suppress(ImportError):  # Module not available when building docs
     from polars.polars import PySeries
 
 if TYPE_CHECKING:
     from polars import DataFrame, Series
-    from polars.dependencies import pandas as pd
+    from polars._dependencies import pandas as pd
     from polars.type_aliases import PolarsDataType
