+++ py-polars/polars/_utils/construction/utils.py
@@ -4,7 +4,7 @@ import sys
 from functools import lru_cache
 from typing import TYPE_CHECKING, Any, Callable, Sequence, get_type_hints
 
-from polars.dependencies import _check_for_pydantic, pydantic
+from polars._dependencies import _check_for_pydantic, pydantic
 
 if TYPE_CHECKING:
     import pandas as pd
