+++ py-polars/polars/dataframe/frame.py
@@ -31,6 +31,21 @@ from typing import (
 
 import polars._reexport as pl
 from polars import functions as F
+from polars._dependencies import (
+    _GREAT_TABLES_AVAILABLE,
+    _HVPLOT_AVAILABLE,
+    _PANDAS_AVAILABLE,
+    _PYARROW_AVAILABLE,
+    _check_for_numpy,
+    _check_for_pandas,
+    _check_for_pyarrow,
+    great_tables,
+    hvplot,
+    import_optional,
+)
+from polars._dependencies import numpy as np
+from polars._dependencies import pandas as pd
+from polars._dependencies import pyarrow as pa
 from polars._utils.construction import (
     arrow_to_pydf,
     dataframe_to_pydf,
@@ -77,21 +92,6 @@ from polars.datatypes import (
     UInt64,
 )
 from polars.datatypes.group import INTEGER_DTYPES
-from polars.dependencies import (
-    _GREAT_TABLES_AVAILABLE,
-    _HVPLOT_AVAILABLE,
-    _PANDAS_AVAILABLE,
-    _PYARROW_AVAILABLE,
-    _check_for_numpy,
-    _check_for_pandas,
-    _check_for_pyarrow,
-    great_tables,
-    hvplot,
-    import_optional,
-)
-from polars.dependencies import numpy as np
-from polars.dependencies import pandas as pd
-from polars.dependencies import pyarrow as pa
 from polars.exceptions import (
     ColumnNotFoundError,
     ModuleUpgradeRequiredError,
