+++ py-polars/polars/io/ipc/functions.py
@@ -5,13 +5,13 @@ from pathlib import Path
 from typing import IO, TYPE_CHECKING, Any, Sequence
 
 import polars._reexport as pl
+from polars._dependencies import import_optional
 from polars._utils.deprecation import deprecate_renamed_parameter
 from polars._utils.various import (
     is_str_sequence,
     normalize_filepath,
 )
 from polars._utils.wrap import wrap_df, wrap_ldf
-from polars.dependencies import import_optional
 from polars.io._utils import (
     is_glob_pattern,
     is_local_file,
