+++ py-polars/polars/sql/context.py
@@ -12,15 +12,15 @@ from typing import (
     overload,
 )
 
+from polars._dependencies import _check_for_pandas, _check_for_pyarrow
+from polars._dependencies import pandas as pd
+from polars._dependencies import pyarrow as pa
 from polars._utils.deprecation import deprecate_renamed_parameter
 from polars._utils.unstable import issue_unstable_warning
 from polars._utils.various import _get_stack_locals
 from polars._utils.wrap import wrap_ldf
 from polars.convert import from_arrow, from_pandas
 from polars.dataframe import DataFrame
-from polars.dependencies import _check_for_pandas, _check_for_pyarrow
-from polars.dependencies import pandas as pd
-from polars.dependencies import pyarrow as pa
 from polars.lazyframe import LazyFrame
 from polars.series import Series
 from polars.type_aliases import FrameType
