+++ py-polars/tests/unit/constructors/test_constructors.py
@@ -13,9 +13,9 @@ import pytest
 from pydantic import BaseModel, Field, TypeAdapter
 
 import polars as pl
+from polars._dependencies import dataclasses, pydantic
 from polars._utils.construction.utils import try_get_type_hints
 from polars.datatypes import numpy_char_code_to_dtype
-from polars.dependencies import dataclasses, pydantic
 from polars.exceptions import ShapeError
 from polars.testing import assert_frame_equal, assert_series_equal
