+++ py-polars/tests/unit/operations/namespaces/temporal/test_add_business_days.py
@@ -11,7 +11,7 @@ import pytest
 from hypothesis import assume, given
 
 import polars as pl
-from polars.dependencies import _ZONEINFO_AVAILABLE
+from polars._dependencies import _ZONEINFO_AVAILABLE
 from polars.exceptions import ComputeError, InvalidOperationError
 from polars.testing import assert_series_equal
