+++ py-polars/tests/unit/test_polars_import.py
@@ -85,7 +85,7 @@ def test_polars_import() -> None:
     ):
         # ensure that we have not broken lazy-loading (numpy, pandas, pyarrow, etc).
         lazy_modules = [
-            dep for dep in pl.dependencies.__all__ if not dep.startswith("_")
+            dep for dep in pl._dependencies.__all__ if not dep.startswith("_")
         ]
         for mod in lazy_modules:
             not_imported = not df_import["import"].str.starts_with(mod).any()
